requirements.txt
flask
pypdf2
numpy
faiss-cpu
sentence-transformers
